requirements.txt
requests
beautifulsoup4
attrs
